Fix small selection canvas bug regarding number format

+++ fire_atlas_editor/lib/screens/editor_screen/widgets/selection_canvas/selection_form.dart
@@ -48,7 +48,7 @@ class _SelectionFormState extends State<SelectionForm> {
       final selection = widget.editingSelection!;
       selectionNameController.text = selection.id;
       frameCountController.text = selection.frameCount.toString();
-      stepTimeController.text = (selection.stepTime * 1000).toString();
+      stepTimeController.text = (selection.stepTime * 1000).toInt().toString();
 
       _selectionType = SelectionType.ANIMATION;
     }
